+++ clang/lib/AST/ItaniumMangle.cpp
@@ -5008,8 +5008,13 @@ recurse:
     break;
   }
 
+  case Expr::BlockExprClass: {
+    NotPrimaryExpr();
+    mangleUnqualifiedBlock(cast<BlockExpr>(E)->getBlockDecl());
+    break;
+  }
+
   // FIXME: invent manglings for all these.
-  case Expr::BlockExprClass:
   case Expr::ChooseExprClass:
   case Expr::CompoundLiteralExprClass:
   case Expr::ExtVectorElementExprClass:
@@ -6779,7 +6784,7 @@ void CXXNameMangler::mangleValueInTemplateArg(QualType T, const APValue &V,
 
   case APValue::LValue: {
     // Proposed in https://github.com/itanium-cxx-abi/cxx-abi/issues/47.
-    assert((T->isPointerOrReferenceType()) &&
+    assert((T->isPointerOrReferenceType() || T->isBlockPointerType()) &&
            "unexpected type for LValue template arg");
 
     if (V.isNullPointer()) {
@@ -6847,7 +6852,7 @@ void CXXNameMangler::mangleValueInTemplateArg(QualType T, const APValue &V,
           Out << "cv";
           mangleType(T);
         }
-        if (T->isPointerType())
+        if (T->isPointerOrBlockPointerType())
           Out << "ad";
         Out << "so";
         mangleType(T->isVoidPointerType()
@@ -6862,7 +6867,7 @@ void CXXNameMangler::mangleValueInTemplateArg(QualType T, const APValue &V,
           Out << "cv";
           mangleType(T);
         }
-        if (T->isPointerType()) {
+        if (T->isPointerOrBlockPointerType()) {
           NotPrimaryExpr();
           Out << "ad";
         }