+++ clang/docs/ReleaseNotes.rst
@@ -192,6 +192,8 @@ What's New in Clang |release|?
 C++ Language Changes
 --------------------
 
+- Clang now supports block pointers as non-type template parameters. (#GH189247)
+
 - ``__is_trivially_equality_comparable`` no longer returns false for all enum types. (#GH132672)
 - ``auto`` parameters are now available in all C++ language modes as an extension.